Fine tuned test cases
