+++ resources/views/components/flash.blade.php
@@ -1,0 +1,9 @@
+@if (session()->has('livewire-alert'))
+    <script> 
+        window.onload = event => {
+            flashAlert(
+                @json(session('livewire-alert'))
+            ) 
+        }
+    </script>
+@endif
