+++ resources/views/components/scripts.blade.php
@@ -3,12 +3,4 @@
     {!! file_get_contents($jsPath) !!}
 </script>
 
-@if (session()->has('livewire-alert'))
-    <script> 
-        window.onload = event => {
-            flashAlert(
-                @json(session('livewire-alert'))
-            ) 
-        }
-    </script>
-@endif
+<x-livewire-alert::flash />
