+++ tests/LivewireAlertTest.php
@@ -4,20 +4,19 @@ namespace Jantinnerezo\LivewireAlert\Tests;
 
 use Jantinnerezo\LivewireAlert\Exceptions\AlertException;
 use Livewire\Livewire;
-use Jantinnerezo\LivewireAlert\LivewireAlert;
 
 class LivewireAlertTest extends TestCase
 {
     public function testBasicAlert(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->call('showAlert')
             ->assertDispatchedBrowserEvent('alert');
     }
 
     public function testBasicFlashAlert(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('flash', true)
             ->call('showAlert')
             ->assertRedirect('/')
@@ -26,43 +25,43 @@ class LivewireAlertTest extends TestCase
 
     public function testAlertConfirm(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('configuration.showConfirmButton', true)
             ->set('configuration.onConfirmed', 'confirmed')
             ->call('showConfirmAlert')
-            ->emit('confirmed')
-            ->assertDispatchedBrowserEvent('alert');
+            ->assertDispatchedBrowserEvent('alert')
+            ->emit('confirmed');
     }
 
     public function testAlertDenied(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('configuration.showDenyButton', true)
             ->set('configuration.onDenied', 'denied')
             ->call('showAlert')
-            ->emit('denied')
-            ->assertDispatchedBrowserEvent('alert');
+            ->assertDispatchedBrowserEvent('alert')
+            ->emit('denied');
     }
 
     public function testAlertDismissed(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('configuration.showCancelButton', true)
             ->set('configuration.onDismissed', 'dismissed')
             ->call('showAlert')
-            ->emit('dismissed')
-            ->assertDispatchedBrowserEvent('alert');
+            ->assertDispatchedBrowserEvent('alert')
+            ->emit('dismissed');
     }
 
     public function testProgressDismissal(): void
     {
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('configuration.timerProgressBar', true)
             ->set('configuration.timer', 3000)
             ->set('configuration.onProgressFinished', 'progressFinished')
             ->call('showAlert')
-            ->emit('progressFinished')
-            ->assertDispatchedBrowserEvent('alert');
+            ->assertDispatchedBrowserEvent('alert')
+            ->emit('progressFinished');
     }
 
     public function testIfExceptionIsThrownWhenIconIsInvalid()
@@ -72,19 +71,7 @@ class LivewireAlertTest extends TestCase
         $this->expectException(AlertException::class);
         $this->expectExceptionMessage("Invalid '{$invalidIcon}' alert icon.");
         
-        Livewire::test(LivewireAlert::class)
-            ->set('status', $invalidIcon)
-            ->call('showAlert');
-    }
-
-    public function testIfExceptionIsThrownWhenConfigurationKeyIsUnknown()
-    {
-        $invalidIcon = 'failed';
-        
-        $this->expectException(AlertException::class);
-        $this->expectExceptionMessage("Invalid '{$invalidIcon}' alert icon.");
-        
-        Livewire::test(LivewireAlert::class)
+        Livewire::test(TestComponent::class)
             ->set('status', $invalidIcon)
             ->call('showAlert');
     }
